npu: close 13 CodeQL integer-overflow alerts (unblocks the last red check)
The "CodeQL" check reports 13 new alerts in code changed by this PR, all one class:

    Multiplication result may overflow 'int' before it is converted to
    'size_t' / 'unsigned long' / 'long'      (and one 'float' -> 'double')

Each fix widens an operand so the arithmetic happens in the destination type.
No expression changes value for any input these call sites actually see; the
point is that the multiply no longer happens in the narrow type first.

  dequant_q4nx.cpp        5x  calloc((*out_rows) * (*out_cols), ...)
                              -> (size_t)(*out_rows) * (size_t)(*out_cols)
                              CodeQL flagged the 2 inside this PR's diff; all 5
                              sites share the identical expression and risk, so
                              all 5 are fixed.
  npu_engine_universal.cpp 2x  ... + lh * HD + d]
                              -> ... + (size_t)lh * HD + d]
                              the surrounding terms were already cast; lh*HD
                              was the one int*int left in a size_t sum.
  bench_fused_rmsnorm_qkv  2x  i < M * H / i < H * N   -> (long)M * H / (long)H * N
  bench_gemm_bf16_analyt.  2x  i < M * K / i < K * N   -> (long)M * K / (long)K * N
  bench_rms_norm.cpp       2x  i < M * H               -> (long)M * H   (both sites)
  attn_kernel_bench.cpp    2x  q[h * HD + i] / v[... kv * HD + i]
                              -> (size_t)h * HD / (size_t)kv * HD
  replay_attn.cpp          1x  s+=a*b; sa+=a*a; sb+=b*b;
                              -> (double)a*b / (double)a*a / (double)b*b
                              (accumulators are double; the products were float)

Verified: the engine TU (`npu_engine_universal.cpp`) and `dequant_q4nx.cpp` both
pass -fsyntax-only clean, as do the three bench tests and attn_kernel_bench.
`replay_attn.cpp` does NOT syntax-check, but it fails identically with and
without this change — a deprecated wstring_convert in a vendored FLM header
under GCC 15, confirmed by stashing the edit and re-running.

+++ engine/npu/src/dequant_q4nx.cpp
@@ -90,7 +90,7 @@ static float* dequant_i8_core(const uint8_t* data, int i8_rows, int in_features,
     *out_rows = n_tile_rows * TILE_ROWS;
     *out_cols = n_tile_cols * tile_cols;
 
-    float* out = static_cast<float*>(std::calloc((*out_rows) * (*out_cols), sizeof(float)));
+    float* out = static_cast<float*>(std::calloc((size_t)(*out_rows) * (size_t)(*out_cols), sizeof(float)));
     if (!out) return nullptr;
 
     const int row_bytes = tile_cols * 20;  // scales(2c) + zeros(2c) + packed(16c)
@@ -149,7 +149,7 @@ extern "C" float* dequant_i8_signed_to_float_ex(const uint8_t* data, int i8_rows
     *out_rows = n_tile_rows * TILE_ROWS;
     *out_cols = n_tile_cols * TILE_COLS;
 
-    float* out = static_cast<float*>(std::calloc((*out_rows) * (*out_cols), sizeof(float)));
+    float* out = static_cast<float*>(std::calloc((size_t)(*out_rows) * (size_t)(*out_cols), sizeof(float)));
     if (!out) return nullptr;
 
     for (int ir = 0; ir < i8_rows; ir++) {
@@ -196,7 +196,7 @@ extern "C" float* dequant_q8_0_to_float_ex(const uint8_t* data, int i8_rows, int
     *out_rows = n_tile_rows * TILE_ROWS;
     *out_cols = n_tile_cols * TILE_COLS;
 
-    float* out = static_cast<float*>(std::calloc((*out_rows) * (*out_cols), sizeof(float)));
+    float* out = static_cast<float*>(std::calloc((size_t)(*out_rows) * (size_t)(*out_cols), sizeof(float)));
     if (!out) return nullptr;
 
     for (int ir = 0; ir < i8_rows; ir++) {
@@ -252,7 +252,7 @@ extern "C" float* dequant_i8_group_signed_to_float_ex(const uint8_t* data, int i
     *out_rows = n_tile_rows * TILE_ROWS;
     *out_cols = n_tile_cols * TILE_COLS;
 
-    float* out = static_cast<float*>(std::calloc((*out_rows) * (*out_cols), sizeof(float)));
+    float* out = static_cast<float*>(std::calloc((size_t)(*out_rows) * (size_t)(*out_cols), sizeof(float)));
     if (!out) return nullptr;
 
     for (int ir = 0; ir < i8_rows; ir++) {
@@ -310,7 +310,7 @@ extern "C" float* dequant_i8_4736_to_float(const uint8_t* data, int i8_rows, int
     *out_rows = n_tile_rows * TILE_ROWS;
     *out_cols = n_tile_cols * TILE_COLS;
 
-    float* out = static_cast<float*>(std::calloc((*out_rows) * (*out_cols), sizeof(float)));
+    float* out = static_cast<float*>(std::calloc((size_t)(*out_rows) * (size_t)(*out_cols), sizeof(float)));
     if (!out) return nullptr;
 
     for (int ir = 0; ir < i8_rows; ir++) {
